Conflicts:
	webserver/src/main/java/controller/LoginController.java
	webserver/src/main/java/controller/RequestController.java
	webserver/src/main/webapp/request.xhtml

+++ webserver/src/main/java/util/SentimentRequestStats.java
@@ -4,23 +4,23 @@ import java.util.Date;
 
 public class SentimentRequestStats {
 
-	private float sentiment;
-	private int tweets;
+	private double sentiment;
+	private long tweets;
 	private double intervalMin;
 	private double intervalMax;
 	private Date from;
 	private Date to;
 	
-	public float getSentiment() {
+	public double getSentiment() {
 		return sentiment;
 	}
-	public void setSentiment(float sentiment) {
+	public void setSentiment(double sentiment) {
 		this.sentiment = sentiment;
 	}
-	public int getTweets() {
+	public long getTweets() {
 		return tweets;
 	}
-	public void setTweets(int tweets) {
+	public void setTweets(long tweets) {
 		this.tweets = tweets;
 	}
 	public double getIntervalMin() {
